+++ bitorch/layers/__init__.py
@@ -1,3 +1,9 @@
+"""
+This submodule contains adapted pytorch layers that use quantization functions on their weights
+and activations before forwarding them. These layers use the quantization functions specified in the
+quantization submodule.
+"""
+
 from .debug_layers import (
     InputGraphicalDebug,
     InputPrintDebug,
@@ -6,8 +12,9 @@ from .debug_layers import (
     ShapePrintDebug
 )
 from .qactivation import QActivation
-from .qconv import QConv1d, QConv2d, QConv3d
-from .qconv_noact import QConv1d_NoAct, QConv2d_NoAct, QConv3d_NoAct
+from .qconv1d import QConv1d, QConv1d_NoAct
+from .qconv2d import QConv2d, QConv2d_NoAct
+from .qconv3d import QConv3d, QConv3d_NoAct
 from .qlinear import QLinear
 from .pact import Pact
 from .qembedding import QEmbedding, QEmbeddingBag